pyproject.toml
[tool.poetry]
name = "bankruptcy-parser"
version = "0.1.0"
description = "Async парсер банкротств с fedresurs.ru и kad.arbitr.ru"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiohttp = "^3.11"
sqlalchemy = {extras = ["asyncio"], version = "^2.0"}
asyncpg = "^0.30"
openpyxl = "^3.1"
tenacity = "^8.3"
python-dotenv = "^1.0"
beautifulsoup4 = "^4.12"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
